+++ MiDAS/R/MiDAS_stats.R
@@ -335,6 +335,29 @@ omnibusTest <- function(object,
 #' analysis to be performed iteratively on groups of variables (like residues at
 #' particular amino acid position) using likelihood ratio test.
 #'
+#' Argument \code{inheritance_model} specifies the inheritance model that should
+#' be applyed to experiment's data. Following choices are available: 
+#' \itemize{
+#'   \item{"dominant"}{
+#'     carrier status is sufficient for expression of the phenotype (non-carrier: 
+#'     0, heterozygous & homozygous carrier: 1).
+#'   }
+#'   \item{"recessive"}{
+#'     two copies are required for expression of the phenotype (non-carrier & 
+#'     heterozygous carrier: 0, homozygous carrier: 1).
+#'   }
+#'   \item{"additive"}{
+#'     allele dosage matters, homozygous carriers show stronger phenotype 
+#'     expression or higher risk than heterozygous carriers (non-carrier = 0, 
+#'     heterozygous carrier = 1, homozygous carrier = 2).
+#'   }
+#'   \item{"overdominant"}{
+#'     heterozygous carriers are at higher risk compared to non-carriers or 
+#'     homozygous carriers (non-carrier & homozygous carrier = 0, heterozygous 
+#'     carrier = 1).
+#'   }
+#' }  
+#'  
 #' \code{correction} specifies p-value adjustment method to use, common choice
 #' is Benjamini & Hochberg (1995) (\code{"BH"}). Internally this is passed to
 #' \link[stats]{p.adjust}.
@@ -434,7 +457,7 @@ runMiDAS <- function(object,
   if (! is.null(inheritance_model)) {
     if (isExperimentInheritanceModelApplicable(object_details$data[[experiment]])) {
       assert_that(
-        characterMatches(inheritance_model, c("dominant", "recessive", "additive"))
+        characterMatches(inheritance_model, c("dominant", "recessive", "additive", "overdominance"))
       )
       object_details$data[[experiment]] <- applyInheritanceModel(
         experiment = object_details$data[[experiment]],
@@ -713,7 +736,7 @@ runMiDAS_linear_omnibus <- function(call,
                          ""
   )
   term_name <- switch (experiment,
-                       "hla_aa" = "residue",
+                       "hla_aa" = "residues",
                        "term"
   )
   results <- results %>%
@@ -722,7 +745,7 @@ runMiDAS_linear_omnibus <- function(call,
     select(
       !! group_name := .data$group,
       !! term_name := .data$term,
-      .data$dof,
+      .data$df,
       .data$statistic,
       .data$p.value,
       .data$p.adjusted
@@ -847,7 +870,7 @@ runMiDAS_conditional_omnibus <- function(call,
                          ""
   )
   term_name <- switch (experiment,
-                       "hla_aa" = "residue",
+                       "hla_aa" = "residues",
                        "term"
   )
 
@@ -858,7 +881,7 @@ runMiDAS_conditional_omnibus <- function(call,
       select(
         !! group_name := .data$group,
         !! term_name := .data$term,
-        .data$dof,
+        .data$df,
         .data$statistic,
         .data$p.value,
         .data$p.adjusted,