+++ MiDAS/R/MiDAS_transformationFunctions.R
@@ -579,8 +579,8 @@ getAAFrequencies <- function(aa_variation) {
 #'   included in the output table. Can be also used to rename selected
 #'   variables, see examples.
 #' @param format String \code{"latex"} or \code{"html"}.
-#' @param header String specifying header for result table. If \code{NULL}
-#'   no header is added.
+#' @param header String specifying header for result table. If \code{NULL} no header is added.
+#' @param scroll_box_height A character string indicating the height of the table.
 #'
 #' @return Character vector of formatted table source code.
 #'
@@ -612,7 +612,8 @@ formatResults <- function(results,
                           arrange_by = "p.value",
                           select_cols = c("term", "estimate", "std.error", "p.value", "p.adjusted"),
                           format = c("html", "latex"),
-                          header = NULL
+                          header = NULL,
+                          scroll_box_height = "400px"
                           ) {
   assert_that(
     is.character(filter_by),
@@ -620,7 +621,8 @@ formatResults <- function(results,
     is.character(select_cols),
     is.string(format),
     stringMatches(format, choice = c("html", "latex")),
-    isStringOrNULL(header)
+    isStringOrNULL(header),
+    is.string(scroll_box_height)
   )
 
   filter_by <- parse_exprs(filter_by)
@@ -648,13 +650,13 @@ formatResults <- function(results,
   }
 
   results %<>%
-    kable(format = format, format.args = list(digits = 4, scientific = -5)) %>%
+    kable(format = format, format.args = list(digits = 4, scientific = -3)) %>%
     add_header_above(header = header)
 
   if (format == "html") {
     results %<>%
       kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
-      scroll_box(width = "100%", height = "200px")
+      scroll_box(width = "100%", height = scroll_box_height)
   }
 
   return(results)
@@ -692,13 +694,15 @@ kableResults <- function(results,
                          colnames = NULL,
                          header = "MiDAS analysis results",
                          pvalue_cutoff = NULL,
-                         format = getOption("knitr.table.format")) {
+                         format = getOption("knitr.table.format"),
+                         scroll_box_height = "400px") {
   assert_that(
     is.data.frame(results),
     isCharacterOrNULL(colnames),
     isNumberOrNULL(pvalue_cutoff),
     is.string(format),
-    stringMatches(format, choice = c("html", "latex"))
+    stringMatches(format, choice = c("html", "latex")),
+    is.string(scroll_box_height)
   )
   if (! is.null(colnames)) {
     assert_that(
@@ -727,7 +731,8 @@ kableResults <- function(results,
       arrange_by = "p.value",
       select_cols = select_cols,
       format = format,
-      header = header
+      header = header,
+      scroll_box_height = scroll_box_height
     )
 
   return(results)
@@ -976,7 +981,6 @@ filterExperimentByFrequency <- function(experiment,
 #' @return Data frame containing variables and their corresponding frequencies.
 #'
 #' @importFrom assertthat assert_that is.string see_if
-#' @importFrom formattable percent
 #' @importFrom SummarizedExperiment assay
 #'
 getExperimentFrequencies <-
@@ -1024,13 +1028,13 @@ getExperimentFrequencies.matrix <-
         left_join(counts_df, ref, by = c("term" = "var"))
     }
 
-    # format frequencies as percent
-    counts_df[, -c(1, 2)] <-
-      rapply(
-        object = counts_df[, -c(1, 2), drop = FALSE],
-        f = function(col) percent(col),
-        how = "replace"
-      )
+#     # format frequencies as percent
+#     counts_df[, -c(1, 2)] <-
+#       rapply(
+#         object = counts_df[, -c(1, 2), drop = FALSE],
+#         f = function(col) percent(col),
+#         how = "replace"
+#       )
 
     return(counts_df)
   }
@@ -1068,7 +1072,8 @@ getExperimentFrequencies.SummarizedExperiment <-
 #' Under \code{"dominant"} model homozygotes and heterozygotes are coded as
 #' \code{1}. In \code{"recessive"} model homozygotes are coded as \code{1} and
 #' other as \code{0}. In \code{"additive"} model homozygotes are coded as
-#' \code{2} and heterozygotes as \code{1}.
+#' \code{2} and heterozygotes as \code{1}. In \code{"overdominance"} homozygotes
+#' (both \code{0} and \code{2}) are coded as \code{0} and heterozygotes as \code{1}.
 #'
 #' @param experiment Matrix or SummarizedExperiment object.
 #' @param inheritance_model String specifying inheritance model to use.
@@ -1079,7 +1084,7 @@ getExperimentFrequencies.SummarizedExperiment <-
 #'
 applyInheritanceModel <-
   function(experiment,
-           inheritance_model = c("dominant", "recessive", "additive")) {
+           inheritance_model = c("dominant", "recessive", "additive", "overdominance")) {
     UseMethod("applyInheritanceModel", experiment)
   }
 
@@ -1087,17 +1092,27 @@ applyInheritanceModel <-
 #' @method applyInheritanceModel matrix
 #'
 applyInheritanceModel.matrix <- function(experiment,
-                                         inheritance_model =  c("dominant", "recessive", "additive")) {
-  .classify <- function(x, val) {
+                                         inheritance_model =  c("dominant", "recessive", "additive", "overdominance")) {
+  .classifyGte <- function(x, val) {
+  # classifies vector as being greater than number
+  # specificly for 1 we classify as dominat, and 2 as recessive
     x <- x >= val
+    mode(x) <- "integer"
+    x
+  }
+  .classifyEq <- function(x, val) {
+  # classify vector as being equal to number
+  # specificly for 1 we classify as overdominance
+    x <- x == val
     mode(x) <- "integer"
     x
   }
   switch (
     inheritance_model,
     "additive" = experiment,
-    "dominant" = .classify(experiment, 1), # ifelse(x >= 1, 1, 0)
-    "recessive" = .classify(experiment, 2) # ifelse(x >= 2, 1, 0)
+    "dominant" = .classifyGte(experiment, 1), # ifelse(x >= 1, 1, 0)
+    "recessive" = .classifyGte(experiment, 2), # ifelse(x >= 2, 1, 0)
+    "overdominance" = .classifyEq(experiment, 1) # ifelse(x == 1, 1, 0)
   )
 }
 
@@ -1105,7 +1120,7 @@ applyInheritanceModel.matrix <- function(experiment,
 #' @method applyInheritanceModel SummarizedExperiment
 #'
 applyInheritanceModel.SummarizedExperiment <- function(experiment,
-                                                       inheritance_model =  c("dominant", "recessive", "additive")) {
+                                                       inheritance_model =  c("dominant", "recessive", "additive", "overdominance")) {
   SummarizedExperiment::assay(experiment) <-
     applyInheritanceModel(SummarizedExperiment::assay(experiment), inheritance_model)